[TS-379] Refactor: habit record page path refactoring
