+++ src/components/Habit/PracticeRecord/PracticeRecord.style.ts
