+++ src/constants/path.ts
@@ -11,6 +11,8 @@ export const PATH = {
 		`/rest-record/${habitId}/${year}/${month}/${date}`,
 	PRACTICE_RECORD: (habitId: string, year: string, month: string, date: string) =>
 		`/practice-record/${habitId}/${year}/${month}/${date}`,
+	HABIT_RECORD: (habitId: number, year: number, month: number, date: number) =>
+		`/habit-record/${habitId}/${year}/${month}/${date}`,
 	EDIT_HABIT: (habitId?: string) => `/habit-edit/${habitId}`,
 	DELETE_HABIT: (habitId?: string) => `/habit-delete/${habitId}`,
 } as const;
