+++ src/pages/HabitRecordPage.tsx
@@ -2,7 +2,7 @@ import FooterBtn from "@/components/common/FooterBtn/FooterBtn";
 import Header from "@/components/common/Header/Header";
 import HabitDetail from "@/components/Habit/HabitDetail/HabitDetail";
 
-function HabitDetailPage() {
+const HabitRecordPage = () => {
 	return (
 		<>
 			<Header>
@@ -12,6 +12,6 @@ function HabitDetailPage() {
 			<FooterBtn text="확인" isTransparent />
 		</>
 	);
-}
+};
 
-export default HabitDetailPage;
+export default HabitRecordPage;
